+++ crypto/crmf/crmf_int.h
@@ -18,7 +18,7 @@
 # include <openssl/err.h>
 
 /* explicit #includes not strictly needed since implied by the above: */
-# include <openssl/ossl_typ.h>
+# include <openssl/types.h>
 # include <openssl/safestack.h>
 # include <openssl/x509.h>
 # include <openssl/x509v3.h>
