+++ crypto/x509/v3_admis.c
@@ -9,7 +9,7 @@
 #include <stdio.h>
 #include "internal/cryptlib.h"
 #include <openssl/conf.h>
-#include <openssl/ossl_typ.h>
+#include <openssl/types.h>
 #include <openssl/asn1.h>
 #include <openssl/asn1t.h>
