+++ crypto/x509/x509_meth.c
@@ -14,7 +14,7 @@
 #include "internal/cryptlib.h"
 #include <openssl/asn1.h>
 #include <openssl/x509.h>
-#include <openssl/ossl_typ.h>
+#include <openssl/types.h>
 #include "x509_lcl.h"
 
 X509_LOOKUP_METHOD *X509_LOOKUP_meth_new(const char *name)
