+++ include/openssl/conf.h
@@ -14,7 +14,7 @@
 # include <openssl/lhash.h>
 # include <openssl/safestack.h>
 # include <openssl/e_os2.h>
-# include <openssl/ossl_typ.h>
+# include <openssl/types.h>
 # include <openssl/conferr.h>
 
 #ifdef  __cplusplus
