+++ include/openssl/crmf.h
@@ -23,7 +23,7 @@
 #  include <openssl/x509v3.h> /* for GENERAL_NAME etc. */
 
 /* explicit #includes not strictly needed since implied by the above: */
-#  include <openssl/ossl_typ.h>
+#  include <openssl/types.h>
 #  include <openssl/x509.h>
 
 #  ifdef  __cplusplus
