+++ include/openssl/modes.h
@@ -11,7 +11,7 @@
 # define OPENSSL_MODES_H
 
 # include <stddef.h>
-# include <openssl/ossl_typ.h>
+# include <openssl/types.h>
 
 # ifdef  __cplusplus
 extern "C" {
