+++ include/openssl/rand.h
@@ -11,7 +11,7 @@
 # define OPENSSL_RAND_H
 
 # include <stdlib.h>
-# include <openssl/ossl_typ.h>
+# include <openssl/types.h>
 # include <openssl/e_os2.h>
 # include <openssl/randerr.h>
