+++ test/shlibloadtest.c
@@ -12,7 +12,7 @@
 #include <stdlib.h>
 #include <openssl/opensslv.h>
 #include <openssl/ssl.h>
-#include <openssl/ossl_typ.h>
+#include <openssl/types.h>
 #include "internal/dso_conf.h"
 
 typedef void DSO;
